refactor: move counterparty selector to its own file
